requirements.txt
bcrypt==4.1.2
beautifulsoup4==4.12.3
bs4==0.0.2
bson==0.5.10
dnspython==2.6.1
email_validator==2.1.1
espy-contact==0.3
fastapi==0.110.0
gunicorn==21.2.0
motor==3.4.0
playwright==1.41.2
psycopg2-binary==2.9.9
pydantic==2.7.1
pymongo==4.7.0
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.9
requests==2.31.0
selenium==4.18.1
sendgrid==6.11.0
SQLAlchemy==2.0.29
sse-starlette==2.1.0
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1